==> scripts/1 data collection/2022.07.08 python - scrape nec vote count/step 1 - get sub election.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import json

# Step 0 - import election type and election code

filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 0 - election type/dat_step0.xlsx"
dat_step0 = pd.read_excel(filepath,dtype=None)
dat_step0_columns = list(dat_election.columns)

dat_step0

# Step 1 - get sub election type

# Step 1 - get sub election type

url = "http://info.nec.go.kr/bizcommon/selectbox/selectbox_getSubElectionTypeJson.json"
electionId = "0000000000"
electionCode = ""
electionType = ""
dat_post = {'electionId': electionId, "electionCode": electionCode,"electionType":electionType}

bodyparts=[]

for index,row in dat_step0.iterrows():
    dat_post["electionCode"]=row["ElectionCode"]
    dat_post["electionType"]=row["ElectionTypeCode"]
    x = requests.post(url,data=dat_post)
    j = json.loads(x.text)
    body = j["jsonResult"]["body"]
    
    [bodyparts.append(list(row)+[item["CODE"],item["NAME"]]) for item in body]

dat_step1 = pd.DataFrame(bodyparts,columns=list(dat_step0.columns)+["SubElectionCode","SubElectionName"])
dat_step1

# save results from step 1
dat_step1.to_excel("../../data/2.0 current through 2022 data/2022 NEC web scraping/step 1 - get sub election/dat_step1.xlsx",index=False,encoding="utf-8")
