==> scripts/1 data collection/2022.07.08 python - scrape nec vote count/step 2.3 - get city - everything else & concat.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import json

# import Step 1 dataframe
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 1 - get sub election/dat_step1.xlsx"
dat_step1 = pd.read_excel(filepath,dtype=None)

# Step 2.3 - get city code for everything else, including 지선 > 3

# Step 2 - city Code by Sg

# everything else, including 지선 > 3
url = "http://info.nec.go.kr/bizcommon/selectbox/selectbox_cityCodeBySgJson_Old.json"
electionId = "0000000000"
subElectionCode = "" # get from the JSON response to Step 1
electionCode = ""   # same electionCode as in Step 1
dat_post = {'electionId': electionId, "subElectionCode": subElectionCode,"electionCode":electionCode}

dat_step2 = pd.DataFrame(columns=list(dat_step1.columns)+["CityCode","CityName"])
dat_step2

dat_step1

# 지선 1,2,3 from dat_step1
dat_step1_subset = dat_step1[dat_step1["ElectionTypeCode"]!=0]   # exclude 재보궐선거
dat_step1_subset = dat_step1_subset[~((dat_step1_subset["ElectionTypeCode"]==4) & (dat_step1["ElectionCode"].astype(int)<20060531))]   # exclude 지선 1,2,3
dat_step1_subset.shape

dat_step1_subset

bodyparts=[]

for index,row in dat_step1_subset.iterrows():
    dat_post["subElectionCode"]=row["SubElectionCode"]
    dat_post["electionCode"]=row["ElectionCode"]
    x = requests.post(url,data=dat_post)
    j = json.loads(x.text)
    body = j["jsonResult"]["body"]
    
    [bodyparts.append(list(row)+[item["CODE"],item["NAME"]]) for item in body]

bodyparts

dat_step2_sub = pd.DataFrame(bodyparts,columns=dat_step2.columns)
dat_step2_sub

dat_step2_sub.shape

len(bodyparts)

dat_step2_sub.to_excel("../../data/2.0 current through 2022 data/2022 NEC web scraping/step 2 - get city/dat_step2_3.xlsx",index=False,encoding="utf-8")

# Concatenate "get city" steps 2.1, 2.2, 2.3

# import Step 2.1
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 2 - get city/dat_step2_1.xlsx"
dat_step2_1 = pd.read_excel(filepath,dtype=None)

# import Step 2.2
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 2 - get city/dat_step2_1.xlsx"
dat_step2_2 = pd.read_excel(filepath,dtype=None)

# append 2.1, 2.2, 2.3
dat_step2 = pd.concat([dat_step2_1,dat_step2_2,dat_step2_sub],ignore_index=True)

dat_step2.shape

dat_step2

# export concatenated "get city" Step 2
dat_step2.to_excel("../../data/2.0 current through 2022 data/2022 NEC web scraping/step 2 - get city/dat_step2.xlsx",index=False,encoding="utf-8")
