==> scripts/1 data collection/2022.07.08 python - scrape nec vote count/step 3.2 - get town - 지선123.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import json

# Step 3.2 - get town code - 지선 1,2,3

# import Step 2.2 - "get city" codes for 지선 1,2,3
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 2 - get city/dat_step2_2.xlsx"
dat_step2_2 = pd.read_excel(filepath,dtype=None)

# "get town" url and post data for 지선 1,2,3 
url = "http://info.nec.go.kr/bizcommon/selectbox/selectbox_townCodeBySgJson_GuOld.json"
electionId = "0000000000"
electionCode = ""   # same election code as in Step 2
subElectionCode = "" # same sub election code as in Step 2
cityCode = ""   # get city code from Step 2
dat_post = {'electionId': electionId, "subElectionCode": subElectionCode,"electionCode":electionCode,"cityCode":cityCode}

dat_step3 = pd.DataFrame(columns=list(dat_step2_2.columns)+["TownCode","TownName"])
dat_step3

dat_step2_2

bodyparts=[]

for index,row in dat_step2_2.iterrows():
    dat_post["subElectionCode"]=row["SubElectionCode"]
    dat_post["electionCode"]=row["ElectionCode"]
    dat_post["cityCode"]=row["CityCode"]
    x = requests.post(url,data=dat_post)
    j = json.loads(x.text)
    body = j["jsonResult"]["body"]
    
    [bodyparts.append(list(row)+[item["CODE"],item["NAME"]]) for item in body]

bodyparts

dat_step3_2 = pd.DataFrame(bodyparts,columns=dat_step3.columns)
dat_step3_2

dat_step3_2.shape

len(bodyparts)

dat_step3_2.to_excel("../../data/2.0 current through 2022 data/2022 NEC web scraping/step 3 - get town/dat_step3_2.xlsx",index=False,encoding="utf-8")
