==> scripts/1 data collection/2022.07.08 python - scrape nec vote count/step 3.4 - check town.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import json

# Step 3.4 - check town code

#### for 구시군장 in 재보궐 (3.1) and 지선 1,2,3 (3.2)
#### for 구시군장 & 비례대표기초의원 in 3.1, 3.2, 3.3

## check 구시군장 in Step 3.1

# import Step 3.1 - "get town" codes for 재보궐선거
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 3 - get town/dat_step3_1.xlsx"
dat_step3_1 = pd.read_excel(filepath,dtype=None)

dat_step3_1[["SubElectionCode","SubElectionName"]].drop_duplicates()

dat_step3_1[dat_step3_1["SubElectionCode"]==4]

dat_step3_1[dat_step3_1["TownCode"].isnull()]

## check 구시군장 in Step 3.2

# import Step 3.2 - "get town" codes for 지선 1,2,3
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 3 - get town/dat_step3_2.xlsx"
dat_step3_2 = pd.read_excel(filepath,dtype=None)

dat_step3_2[["SubElectionCode","SubElectionName"]].drop_duplicates()

dat_step3_2[dat_step3_2["SubElectionCode"]==4]

dat_step3_2[dat_step3_2["TownName"].isnull()]

## check 구시군장, 비례대표기초의원, 교육감의원 in Step 3.3

# import Step 3.3 - "get town" codes for everything else
filepath = "../../data/2.0 current through 2022 data/2022 NEC web scraping/step 3 - get town/dat_step3_3.xlsx"
dat_step3_3 = pd.read_excel(filepath,dtype=None)

dat_step3_3[["SubElectionCode","SubElectionName"]].drop_duplicates()

dat_step3_3[dat_step3_3["SubElectionCode"]==4] # 구시군장

dat_step3_3[dat_step3_3["SubElectionCode"]==9] # 기초의원 비례대표

dat_step3_3[dat_step3_3["SubElectionCode"]==10] # 교육의원선거
