==> scripts/1 data collection/2022.07.08 python - scrape nec vote count/step 4 - get sggCity sggTown Excel - 11edboss.ipynb <==
import winsound
frequency = 1000  # Set Frequency To 2500 Hertz
duration = 1000  # Set Duration To 1000 ms == 1 second

#winsound.Beep(frequency, duration)

def success(frequency,duration,n):
    for i in range(1, n):
        winsound.Beep(i * frequency, duration)
        
def fail(frequency,duration,n):
    for i in range(1,n):
        winsound.Beep((n-i) * frequency, duration)
        
success(500,500,6)
fail(500,500,6)

import requests
from bs4 import BeautifulSoup as bs
import pandas as pd
import json
import numpy as np
import lxml

# Step 4 - get sggCity, sggTown, and Excel - subelection 11 (교육감)

## Step 4.0 - choose subelection

subelection_code = 11
subelelection_name = "11edboss"

dat_step3 = pd.read_excel("../../data_electoral/2.0 current through 2022 data/2022 NEC web scraping/step 3 - get town/dat_step3.xlsx")

dat_step3[["SubElectionCode","SubElectionName"]].drop_duplicates().sort_values(by="SubElectionCode")

dat_step3

## Step 4.1 - get sggCityCode - for subelection = 1,2,3,7: sggCityCode = NaN

## Step 4.2 - get sggTownCode - for subelection = 1,2,3,7: sggTownCode = NaN

dat_step4 = dat_step3[dat_step3["SubElectionCode"]==subelection_code].reset_index()

dat_step4["SggCityCode"]=np.nan
dat_step4["SggCityName"]=np.nan
dat_step4["SggTownCode"]=np.nan
dat_step4["SggTownName"]=np.nan

dat_step4

dat_step4.shape

dat_step4.to_excel("../../data_electoral/2.0 current through 2022 data/2022 NEC web scraping/step 4 - get sggCity & sggTown/dat_step4_"+subelelection_name+".xlsx",index=False,encoding="utf-8")

## Step 4.3 - download 개표현황(읍면동) Excel files

dat_step4 = pd.read_excel("../../data_electoral/2.0 current through 2022 data/2022 NEC web scraping/step 4 - get sggCity & sggTown/dat_step4_"+subelelection_name+".xlsx")

dat_step4

# query electionInfo page - for subelection = 1 (president)

url = "http://info.nec.go.kr/electioninfo/electionInfo_report.xhtml"

electionId = "0000000000"
requestURI = "/WEB-INF/jsp/electioninfo/0000000000/vc/vccp04.jsp"
topMenuId = "VC"
secondMenuId = "VCCP04"
menuId = "VCCP04"
statementId = "VCCP04_#2"
oldElectionType = "1"      # fixed - for 지선 >3 = 1 (and for 지선 1,2,3 = 0, but 교육감 appears 지선 5)
electionType = "4"         # ElectionTypeCode - variable - 재보궐 = 0, 지선 = 4, 
electionName = "20170509"  # ElectionCode - variable
electionCode = "11"         # SubElectionCode - fixed - 교육감 = 11
cityCode = "2600"          # CityCode - variable
sggCityCode = "-1"         # SggCityCode - NaN
townCodeFromSgg = "-1"     # SggTownCode - NaN
townCode = "2604"          # TownCode - variable
#sggTownCode = ""          # SGGTownCode - not a parameter
x = "62"
y = "20"

dat_post = {'electionId': electionId,
            "requestURI": requestURI,
            "topMenuId": topMenuId,
            "secondMenuId": secondMenuId,
            "menuId":menuId,
            "statementId":statementId,
            "oldElectionType":oldElectionType,
            "electionType":electionType,       # ElectionTypeCode - variable
            "electionName":electionName,       # ElectionCode - variable  
            "electionCode":electionCode,       # SubElectionCode - fixed
            "cityCode":cityCode,               # CityCode - variable
            "sggCityCode":sggCityCode,         # SggCityCode - fixed
            "townCodeFromSgg":townCodeFromSgg, # TownCode - fixed
            "townCode":townCode,               # TownCode - variable
            #"sggTownCode":"",                 # SggTownCode - omitted from subelection = 1 (president)
            "x":x,
            "y":y}

## test requests
d = dat_step4
# d = dat_step3[dat_step3["ElectionTypeCode"]==4]
# d = d[d["ElectionCode"]>=20060531]
# d

excelpath = "../../data_electoral/2.0 current through 2022 data/2022 NEC web scraping/step 4 - get sggCity & sggTown/votecount_"+subelelection_name+"/"

try:

    for index,row in d[337:800].iterrows():

        # post parameters
        
        # variable - for 지선 1,2,3 = 0; for 지선 >3 = 1
        if row["ElectionTypeCode"]==4 and row["ElectionCode"]<20060531:
            dat_post["oldElectionType"]= "0"      
        else:
            dat_post["oldElectionType"]= "1"      

        dat_post["electionType"]=row["ElectionTypeCode"]
        dat_post["electionName"]=row["ElectionCode"]
        #dat_post["electionCode"]=row["SubElectionCode"]
        dat_post["cityCode"]=row["CityCode"]
        #dat_post["sggCityCode"]=row["SggCityCode"]
        #dat_post["townCodeFromSgg"]=row["TownCode"]
        dat_post["townCode"]=row["TownCode"]
        #dat_post["sggTownCode"]=row["SggTownCode"]
                
        x = requests.post(url,data=dat_post)
        #print(row)
        print(index,row["ElectionTypeName"],row["ElectionName"],row["ElectionCode"],row["CityCode"],row["CityName"],row["TownCode"],row["TownName"],x.status_code)

        #j = json.loads(x.text)

        # write html file
        # fo = open("out.html", "w",encoding="utf-8")
        # fo.write(x.text)
        # fo.close()


        # convert html table to pandas and write as an excel file
        tables = pd.read_html(x.text)
        t = tables[0]

        excelfilename = "votecount_"+subelelection_name+"_"+str(row["ElectionCode"])+"_"+str(row["CityCode"])+"_"+str(row["TownCode"])+".xlsx"

        t.to_excel(excelpath+excelfilename)

    success(500,500,6)
    
except:
    
    fail(500,500,6)
    print("failed at",index)
    

# print out last item as an html file
# fo = open("out.html", "w",encoding="utf-8")
# fo.write(x.text)
# fo.close()

## Step 4.4 - stats

d.groupby(["ElectionTypeCode"]).agg({"TownCode":"count"})

d.groupby(["ElectionTypeCode","SubElectionCode"]).agg({"TownCode":"count"})

d.groupby(["ElectionTypeCode","ElectionCode","SubElectionCode"]).agg({"TownCode":"count"})

d.groupby(["ElectionTypeCode","ElectionCode","SubElectionCode"]).agg({"CityCode":"nunique"})
